FileOffset is now uint64_t if available. Also added support for 64 bit offsets in FileWriter and FileScanner for Linux.

+++ libosmscout-import/src/osmscout/import/GenCityStreetIndex.cpp
@@ -510,7 +510,7 @@ namespace osmscout {
       for (std::list<LocRef>::const_iterator o=n->second.begin();
            o!=n->second.end();
            ++o) {
-        if (!writer.WriteNumber(o->reference.type)) {
+        if (!writer.WriteNumber((uint32_t)o->reference.type)) {
           return false;
         }
